migrate to go-flags, export Command for embedding
